Added new Job interface changes to partial files and bumped version to 1.1. Started first round of tests.
